requirements.txt
gradio==4.44.0
PyMuPDF==1.23.26
pylibdmtx==0.1.10
Pillow==10.4.0
uvicorn[standard]==0.30.6
gradio>=4.0.0
PyMuPDF>=1.23.0
pylibdmtx>=0.1.10
Pillow>=10.0.0
numpy>=1.24.0
uvicorn>=0.24.0
fastapi>=0.104.0
